+++ modules/mapper-extras/src/main/java/org/elasticsearch/index/mapper/RankFeatureFieldMapper.java
@@ -40,7 +40,7 @@ import java.util.function.Supplier;
 /**
  * A {@link FieldMapper} that exposes Lucene's {@link FeatureField}.
  */
-public class RankFeatureFieldMapper extends ParametrizedFieldMapper {
+public class RankFeatureFieldMapper extends FieldMapper {
 
     public static final String CONTENT_TYPE = "rank_feature";
 
@@ -59,7 +59,7 @@ public class RankFeatureFieldMapper extends ParametrizedFieldMapper {
         return ((RankFeatureFieldMapper)in).fieldType();
     }
 
-    public static class Builder extends ParametrizedFieldMapper.Builder {
+    public static class Builder extends FieldMapper.Builder {
 
         private final Parameter<Boolean> positiveScoreImpact
             = Parameter.boolParam("positive_score_impact", false, m -> ft(m).positiveScoreImpact, true);
@@ -141,11 +141,6 @@ public class RankFeatureFieldMapper extends ParametrizedFieldMapper {
     }
 
     @Override
-    protected RankFeatureFieldMapper clone() {
-        return (RankFeatureFieldMapper) super.clone();
-    }
-
-    @Override
     public RankFeatureFieldType fieldType() {
         return (RankFeatureFieldType) super.fieldType();
     }
@@ -189,7 +184,7 @@ public class RankFeatureFieldMapper extends ParametrizedFieldMapper {
     }
 
     @Override
-    public ParametrizedFieldMapper.Builder getMergeBuilder() {
+    public FieldMapper.Builder getMergeBuilder() {
         return new Builder(simpleName()).init(this);
     }
 }
