+++ server/src/main/java/org/elasticsearch/index/mapper/RangeType.java
@@ -707,7 +707,7 @@ public enum RangeType {
                                        boolean includeFrom, boolean includeTo);
 
     public final Mapper.TypeParser parser() {
-        return new ParametrizedFieldMapper.TypeParser((n, c) -> new RangeFieldMapper.Builder(n, this, c.getSettings()));
+        return new FieldMapper.TypeParser((n, c) -> new RangeFieldMapper.Builder(n, this, c.getSettings()));
     }
 
     public final String name;
