+++ server/src/main/java/org/elasticsearch/index/mapper/SourceFieldMapper.java
@@ -203,7 +203,7 @@ public class SourceFieldMapper extends MetadataFieldMapper {
     }
 
     @Override
-    public ParametrizedFieldMapper.Builder getMergeBuilder() {
+    public FieldMapper.Builder getMergeBuilder() {
         return new Builder().init(this);
     }
 }
