+++ server/src/test/java/org/elasticsearch/index/mapper/FakeStringFieldMapper.java
@@ -30,7 +30,7 @@ import java.util.Collections;
 import java.util.List;
 
 // Like a String mapper but with very few options. We just use it to test if highlighting on a custom string mapped field works as expected.
-public class FakeStringFieldMapper extends ParametrizedFieldMapper {
+public class FakeStringFieldMapper extends FieldMapper {
 
     public static final String CONTENT_TYPE = "fake_string";
 
@@ -41,7 +41,7 @@ public class FakeStringFieldMapper extends ParametrizedFieldMapper {
         FIELD_TYPE.setIndexOptions(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS);
     }
 
-    public static class Builder extends ParametrizedFieldMapper.Builder {
+    public static class Builder extends FieldMapper.Builder {
 
         public Builder(String name) {
             super(name);
@@ -109,7 +109,7 @@ public class FakeStringFieldMapper extends ParametrizedFieldMapper {
     }
 
     @Override
-    public ParametrizedFieldMapper.Builder getMergeBuilder() {
+    public FieldMapper.Builder getMergeBuilder() {
         return new Builder(simpleName()).init(this);
     }
 }
