+++ server/src/test/java/org/elasticsearch/index/mapper/ParametrizedMapperTests.java
@@ -32,7 +32,7 @@ import org.elasticsearch.common.xcontent.json.JsonXContent;
 import org.elasticsearch.index.analysis.AnalyzerScope;
 import org.elasticsearch.index.analysis.IndexAnalyzers;
 import org.elasticsearch.index.analysis.NamedAnalyzer;
-import org.elasticsearch.index.mapper.ParametrizedFieldMapper.Parameter;
+import org.elasticsearch.index.mapper.FieldMapper.Parameter;
 import org.elasticsearch.plugins.MapperPlugin;
 import org.elasticsearch.plugins.Plugin;
 
@@ -88,7 +88,7 @@ public class ParametrizedMapperTests extends MapperServiceTestCase {
         return (TestMapper) in;
     }
 
-    public static class Builder extends ParametrizedFieldMapper.Builder {
+    public static class Builder extends FieldMapper.Builder {
 
         final Parameter<Boolean> fixed
             = Parameter.boolParam("fixed", false, m -> toType(m).fixed, true);
@@ -138,7 +138,7 @@ public class ParametrizedMapperTests extends MapperServiceTestCase {
         }
 
         @Override
-        public ParametrizedFieldMapper build(Mapper.BuilderContext context) {
+        public FieldMapper build(Mapper.BuilderContext context) {
             return new TestMapper(name(), buildFullName(context),
                 multiFieldsBuilder.build(this, context), copyTo.build(), this);
         }
@@ -154,7 +154,7 @@ public class ParametrizedMapperTests extends MapperServiceTestCase {
         }
     }
 
-    public static class TestMapper extends ParametrizedFieldMapper {
+    public static class TestMapper extends FieldMapper {
 
         private final boolean fixed;
         private final boolean fixed2;
