+++ server/src/test/java/org/elasticsearch/index/mapper/TextFieldMapperTests.java
@@ -30,7 +30,6 @@ import org.apache.lucene.analysis.core.WhitespaceAnalyzer;
 import org.apache.lucene.analysis.en.EnglishAnalyzer;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;
-import org.apache.lucene.document.FieldType;
 import org.apache.lucene.index.DocValuesType;
 import org.apache.lucene.index.IndexOptions;
 import org.apache.lucene.index.IndexableField;
@@ -576,7 +575,9 @@ public class TextFieldMapperTests extends MapperTestCase {
             );
             FieldMapper prefix = (FieldMapper) mapper.mappers().getMapper("field._index_prefix");
             assertEquals(prefix.name(), "field._index_prefix");
-            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS_AND_OFFSETS, prefix.fieldType.indexOptions());
+            ParsedDocument doc = mapper.parse(source(b -> b.field("field", "some text")));
+            IndexableField field = doc.rootDoc().getField("field._index_prefix");
+            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS_AND_OFFSETS, field.fieldType().indexOptions());
         }
 
         {
@@ -590,10 +591,11 @@ public class TextFieldMapperTests extends MapperTestCase {
                 )
             );
             FieldMapper prefix = (FieldMapper) mapper.mappers().getMapper("field._index_prefix");
-            FieldType ft = prefix.fieldType;
+            ParsedDocument doc = mapper.parse(source(b -> b.field("field", "some text")));
+            IndexableField field = doc.rootDoc().getField("field._index_prefix");
             assertEquals(prefix.name(), "field._index_prefix");
-            assertEquals(IndexOptions.DOCS, ft.indexOptions());
-            assertFalse(ft.storeTermVectors());
+            assertEquals(IndexOptions.DOCS, field.fieldType().indexOptions());
+            assertFalse(field.fieldType().storeTermVectors());
         }
 
         {
@@ -607,10 +609,11 @@ public class TextFieldMapperTests extends MapperTestCase {
                 )
             );
             FieldMapper prefix = (FieldMapper) mapper.mappers().getMapper("field._index_prefix");
-            FieldType ft = prefix.fieldType;
+            ParsedDocument doc = mapper.parse(source(b -> b.field("field", "some text")));
+            IndexableField field = doc.rootDoc().getField("field._index_prefix");
             assertEquals(prefix.name(), "field._index_prefix");
-            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, ft.indexOptions());
-            assertFalse(ft.storeTermVectors());
+            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, field.fieldType().indexOptions());
+            assertFalse(field.fieldType().storeTermVectors());
         }
 
         {
@@ -624,10 +627,11 @@ public class TextFieldMapperTests extends MapperTestCase {
                 )
             );
             FieldMapper prefix = (FieldMapper) mapper.mappers().getMapper("field._index_prefix");
-            FieldType ft = prefix.fieldType;
+            ParsedDocument doc = mapper.parse(source(b -> b.field("field", "some text")));
+            IndexableField field = doc.rootDoc().getField("field._index_prefix");
             assertEquals(prefix.name(), "field._index_prefix");
-            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, ft.indexOptions());
-            assertTrue(ft.storeTermVectorOffsets());
+            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, field.fieldType().indexOptions());
+            assertTrue(field.fieldType().storeTermVectorOffsets());
         }
 
         {
@@ -641,10 +645,11 @@ public class TextFieldMapperTests extends MapperTestCase {
                 )
             );
             FieldMapper prefix = (FieldMapper) mapper.mappers().getMapper("field._index_prefix");
-            FieldType ft = prefix.fieldType;
+            ParsedDocument doc = mapper.parse(source(b -> b.field("field", "some text")));
+            IndexableField field = doc.rootDoc().getField("field._index_prefix");
             assertEquals(prefix.name(), "field._index_prefix");
-            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, ft.indexOptions());
-            assertFalse(ft.storeTermVectorOffsets());
+            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, field.fieldType().indexOptions());
+            assertFalse(field.fieldType().storeTermVectorOffsets());
         }
     }
 
@@ -664,10 +669,10 @@ public class TextFieldMapperTests extends MapperTestCase {
             assertThat(textField, instanceOf(TextFieldType.class));
             MappedFieldType prefix = ((TextFieldType) textField).getPrefixFieldType();
             assertEquals(prefix.name(), "object.field._index_prefix");
-            FieldMapper mapper
-                = (FieldMapper) mapperService.documentMapper().mappers().getMapper("object.field._index_prefix");
-            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, mapper.fieldType.indexOptions());
-            assertFalse(mapper.fieldType.storeTermVectorOffsets());
+            ParsedDocument doc = mapperService.documentMapper().parse(source(b -> b.field("object.field", "some text")));
+            IndexableField field = doc.rootDoc().getField("object.field._index_prefix");
+            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, field.fieldType().indexOptions());
+            assertFalse(field.fieldType().storeTermVectorOffsets());
         }
 
         {
@@ -685,10 +690,11 @@ public class TextFieldMapperTests extends MapperTestCase {
             assertThat(textField, instanceOf(TextFieldType.class));
             MappedFieldType prefix = ((TextFieldType) textField).getPrefixFieldType();
             assertEquals(prefix.name(), "body.with_prefix._index_prefix");
-            FieldMapper mapper
-                = (FieldMapper) mapperService.documentMapper().mappers().getMapper("body.with_prefix._index_prefix");
-            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, mapper.fieldType.indexOptions());
-            assertFalse(mapper.fieldType.storeTermVectorOffsets());
+            ParsedDocument doc = mapperService.documentMapper().parse(source(b -> b.field("body", "some text")));
+            IndexableField field = doc.rootDoc().getField("body.with_prefix._index_prefix");
+
+            assertEquals(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS, field.fieldType().indexOptions());
+            assertFalse(field.fieldType().storeTermVectorOffsets());
         }
     }
