+++ test/framework/src/main/java/org/elasticsearch/index/mapper/MockFieldMapper.java
@@ -28,7 +28,7 @@ import java.util.Collections;
 import java.util.List;
 
 // this sucks how much must be overridden just do get a dummy field mapper...
-public class MockFieldMapper extends ParametrizedFieldMapper {
+public class MockFieldMapper extends FieldMapper {
     static Settings DEFAULT_SETTINGS = Settings.builder()
         .put(IndexMetadata.SETTING_VERSION_CREATED, Version.CURRENT.id)
         .build();
@@ -50,7 +50,7 @@ public class MockFieldMapper extends ParametrizedFieldMapper {
     }
 
     @Override
-    public ParametrizedFieldMapper.Builder getMergeBuilder() {
+    public FieldMapper.Builder getMergeBuilder() {
         return new Builder(simpleName());
     }
 
@@ -84,7 +84,7 @@ public class MockFieldMapper extends ParametrizedFieldMapper {
     protected void parseCreateField(ParseContext context) {
     }
 
-    public static class Builder extends ParametrizedFieldMapper.Builder {
+    public static class Builder extends FieldMapper.Builder {
         private final MappedFieldType fieldType;
 
         protected Builder(String name) {
