+++ x-pack/plugin/mapper-constant-keyword/src/main/java/org/elasticsearch/xpack/constantkeyword/mapper/ConstantKeywordFieldMapper.java
@@ -32,7 +32,6 @@ import org.elasticsearch.index.mapper.MappedFieldType;
 import org.elasticsearch.index.mapper.Mapper;
 import org.elasticsearch.index.mapper.MapperParsingException;
 import org.elasticsearch.index.mapper.MapperService;
-import org.elasticsearch.index.mapper.ParametrizedFieldMapper;
 import org.elasticsearch.index.mapper.ParseContext;
 import org.elasticsearch.index.mapper.ValueFetcher;
 import org.elasticsearch.index.query.QueryShardContext;
@@ -51,7 +50,7 @@ import java.util.function.Supplier;
 /**
  * A {@link FieldMapper} that assigns every document the same value.
  */
-public class ConstantKeywordFieldMapper extends ParametrizedFieldMapper {
+public class ConstantKeywordFieldMapper extends FieldMapper {
 
     public static final String CONTENT_TYPE = "constant_keyword";
 
@@ -60,11 +59,11 @@ public class ConstantKeywordFieldMapper extends ParametrizedFieldMapper {
     }
 
     @Override
-    public ParametrizedFieldMapper.Builder getMergeBuilder() {
+    public FieldMapper.Builder getMergeBuilder() {
         return new Builder(simpleName()).init(this);
     }
 
-    public static class Builder extends ParametrizedFieldMapper.Builder {
+    public static class Builder extends FieldMapper.Builder {
 
         // This is defined as updateable because it can be updated once, from [null] to any value,
         // by a dynamic mapping update.  Once it has been set, however, the value cannot be changed.
@@ -225,11 +224,6 @@ public class ConstantKeywordFieldMapper extends ParametrizedFieldMapper {
 
     ConstantKeywordFieldMapper(String simpleName, MappedFieldType mappedFieldType) {
         super(simpleName, mappedFieldType, MultiFields.empty(), CopyTo.empty());
-    }
-
-    @Override
-    protected ConstantKeywordFieldMapper clone() {
-        return (ConstantKeywordFieldMapper) super.clone();
     }
 
     @Override
